[BUGFIX] Fix native datetime transformation in Extbase query constraints
Convert DateTime objects to string representation instead of an integer
timestamp for native datetime fields.
Resolve and pass the `ColumnMap` information for any invocation of
`DataMapper->getPlainValue` in order to apply the detection for the
underlying database implementation based on TCA definition of
the respective database field of extbase datetime properties.

Resolves: #63482
Resolves: #79473
Releases: main, 13.4
Change-Id: Ifa842df8728845abee2f922317d444955a46be73
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88840
Reviewed-by: Andreas Kienast <akienast@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benni Mack <benni@typo3.org>