[BUGFIX] Fix f:form SelectViewHelper with wrong localized UID scheme
Extbase `<f:form>` editing with relations to localized records
utilizes a `[defaultLanguageRecordUid]_[localizedRecordUid]` identity
scheme.

The two methods `SelectViewHelper->getOptionValueScalar()` and
`AbstractFormFieldViewHelper->convertToPlainValue()` access
these records with a compound identity like this, and return
their full identity string.

This however fails, when for example a model's property with
persisted localized records is checked against matching these
identities, and where just the defaultLanguageRecordUid is
expected (for an "is selected" check of <option> tags).

With this patch, the `getUid()` method calls are preferred
over the full identity, in case we're dealing with
`DomainObjectInterface` objects.

Tests to validate this behaviour in #106493
are adopted, now showing properly selected categories.

An Important.rest has been created to notify consumers of
the `AbstractFormFieldViewHelper` or
`persistenceManager->getIdentifierByObject()` about possibly
needed changes.

Resolves: #106494
Related: #106493
Related: #87915
Related: #105319
Related: #59992
Releases: main, 13.4
Change-Id: I90dd06f2c8b9fe4a25070081e76fcd49f55adf20
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88891
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>