[BUGFIX] Use "allowed" option for TCA type=group in Extbase
Extbase needed to have a "foreign_table" property in
TCA type=group records, in order to deal with
relations. This has been an issue since many versions,
and is now fixed by using the first entry of "allowed",
which is the documented setting for type=group.

This means, that Extbase developers do not need
to set "foreign_table" on TCA type=group anymore!

Resolves: #95797
Resolves: #62277
Resolves: #48701
Releases: main, 13.4
Change-Id: I381ac4c3aee03475d9567d50763541384d935e7a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88898
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Stefan Froemken <froemken@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>