[BUGFIX] Fix extbase allowProperty() not allowing int keys
With #101292 the introduced types change the signature
of PropertyMappingConfiguration->allowProperties() to:

```
  public function allowProperties(string ...$propertyNames): self
```

In case an extbase f:form though submits this:

```
<input type="hidden"
       name="tx_some_model[__trustedProperties]"
       value="{&quot;item&quot;:{&quot;title&quot;:1,&quot;
         categories&quot;:[1,1,1],&quot;date&quot;:1},
         &quot;submit&quot;:1}9919...">
```

Here the `[1,1,1]` gets passed to `allowProperties()` as int values,
leading to a fatal PHP TypeError.

This means the signature also needs to accept at least ints,
which is performed with this patch.

Note that currently, there are no functional extbase tests in place
that cover most of these type changes affecting f:form handling.
Other issues may also show up with "live" extensions and need
evaluation.

The proof this is broken and can be fixed is available in
patch #106493.

Resolves: #106509
Related: #106493
Related: #101292
Releases: main
Change-Id: I65aeed4d80b7c99799676bccd582cc381b5e9360
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88942
Tested-by: core-ci <typo3@b13.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>