[BUGFIX] Allow `DEFAULT NULL` for `VARCHAR`
The database schema generated by `DefaultTcaSchema` lacks the ability to
generate `VARCHAR` fields with `NULL` as their default value.

The code is adjusted to generate `DEFAULT NULL` for input fields that do
not exceed the maximum length of 255 characters.

Since this code adjusts existing database schema with potentially
negative impact, this patch is aimed for main only.

Resolves: #106656
Releases: main
Change-Id: Icba421059819ddd6cfdb7bf55c3b80d17dc0b7bc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89299
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>