[TASK] Migrate BackendUtility to Schema API
This migrates the last occurrences of $GLOBALS['TCA']
in BackendUtility to proper usages of Schema API.

This temporarily extends `getProcessedValue()` to
accept the columns' TCA configuration. It is required
to support volatile column configurations for FlexForm
fields in `FlexFormValueFormatter`. This workaround
will be removed once FormEngine has been migrated,
as this will then support some kind of mutable TCA schema.

Additionally, a couple of unit tests are migrated
to functionals to allow proper usage of Schema API.

Resolves: #107166
Related: #106007
Releases: main
Change-Id: I74d983f838eceba16483ffbf0ac413f8b8d1d04e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/90190
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>