[BUGFIX] Properly resolve "english" XLIFF translation domain
The implemented language handling in #93334 contained
a flaw in language file resolving sort order.

Given an extension that provides these label files:

- de.locallang.xlf
- en.locallang.xlf
- locallang.xlf

The ordering in `LabelFileResolver->getOrderedFileResources()` would
put all three files in the array key 'en'. Later loops over that
result would regard the last entry as the proper 'en' base language
entry (forfeiting all previous iterations).

If an extension would now provide a file:

- ru.locallang.xlf

this would insert a translated file as the last resolved file into
the array, using this translation as the base-level entry.

So, every locale which has a letter after "l" would be regarded as
the default language, instead of english.

This patch provides a test that would fail with current behavior,
and adjusts the `getOrderedFileResources()` method to actually
only "store" english files in the 'en' scope.

Resolves: #108362
Related: #93334
Releases: main
Change-Id: I852f8f7f4348cbdf8d007f9a2c2db9e85f2c4b40
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91942
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Sebastian Fischer <typo3@evoweb.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>